Move Minimum Spanning Tree Algorithm to its own module (#624)
* refact: move minimum spanning tree algo to its own module

* refact: move min_spanning_tree benches to a different test file
